pyproject.toml
[project]
name = "vetra-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "alembic>=1.18.4",
    "cryptography>=46.0.7",
    "fastapi[standard]>=0.136.0",
    "passlib[argon2]>=1.7.4",
    "pydantic-settings>=2.14.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy[asyncio]>=2.0.49",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
    "taskipy>=1.14.1",
    "types-passlib>=1.7.7.20260211",
    "types-python-jose>=3.5.0.20260408",
]


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pyrefly]
project-includes = ["src/**/*.py*", "src/**/*.ipynb", "main.py"]
project-excludes = ["tests/**/*", "venv/**/*", ".venv/**/*", "migrations/**/*"]
search-path = ["."]

[tool.pyrefly.modules]
fastapi = "typed"
pydantic = "typed"
sqlmodel = "typed"
uvicorn = "stubbed"

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
dev = 'fastapi dev src/main.py'
